fix some issue
