+++ components/Layout/Header/index.js
@@ -193,6 +193,7 @@ export default function Header({
           >
             <Link href="/faucet">{t('menu.developers.faucet')}</Link>
             <Link href="/services/send">Send Payment</Link>
+            <Link href="/services/check">Issue Check</Link>
             <Link href="/services/nft-mint">{t('menu.services.nft-mint')}</Link>
             <Link href="/username">{t('menu.usernames')}</Link>
             <Link href="/xrp-xah-taxes">{t('menu.services.tax-reports')}</Link>
